ARIA-143 Better handle exceptions in the process executor
Previously, if an exception was raised during the starting of a task,
the task's process was permenantly blocked on receiving a message.

The reason was that the execption caused the 'listener thread' to
not send a response the to the task's process, as the exception was not handled inside the 'with' block of the listener thread.

The first change I introduced was to wrap the yielding of the message and
the response inside a try-except-finally block, so the exception will be
handled within the 'with' scope, and to ensure a response is sent to the
task's process.

The second change is to move the sending of the 'task started' message in
the task's process to a place where encountering an exception will be
handled via sending a 'task failed' message back to the listener thread.